dev_sistema_escolar_api/requirements.txt
Django==5.0.2
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.2
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0